+++ front/app/containers/Admin/flexible-pages/NavbarItemRow/index.tsx
@@ -47,6 +47,7 @@ interface Props {
   showAddButton?: boolean;
   addButtonDisabled?: boolean;
   showRemoveButton?: boolean;
+  showDeleteButton?: boolean;
   showViewButton?: boolean;
   viewButtonLink?: string;
   onClickEditButton?: () => void;
@@ -62,6 +63,7 @@ export default ({
   showAddButton,
   addButtonDisabled,
   showRemoveButton,
+  showDeleteButton,
   viewButtonLink,
   onClickEditButton,
   onClickAddButton,
@@ -120,7 +122,7 @@ export default ({
           </Link>
         )}
 
-        {!isDefaultPage && (
+        {!isDefaultPage && showDeleteButton && (
           <Button
             buttonStyle="secondary"
             icon="delete"