+++ front/app/containers/Admin/routes.tsx
@@ -8,6 +8,7 @@ import createAdminUsersRoutes from './users/routes';
 import invitationsRoutes from './invitations/routes';
 import createAdminProjectsRoutes from './projects/routes';
 import settingsRoutes from './settings/routes';
+import flexiblePageRoutes from './flexible-pages/routes';
 import pagesRoutes from './pages/routes';
 import createAdminMessagingRoutes from './messaging/routes';
 import ideasRoutes from './ideas/routes';
@@ -112,6 +113,7 @@ const createAdminRoutes = () => {
       createAdminInitiativesRoutes(),
       createAdminUsersRoutes(),
       createAdminProjectsRoutes(),
+      flexiblePageRoutes(),
       settingsRoutes(),
       pagesRoutes(),
       invitationsRoutes(),