+++ front/app/modules/commercial/customizable_navbar/admin/components/PoliciesSubtitle.tsx
@@ -11,7 +11,7 @@ import { FormattedMessage } from 'utils/cl-intl';
 import messages from './messages';
 
 // utils
-import { NAVIGATION_PATH } from '../containers';
+import { NAVIGATION_PATH } from 'containers/Admin/flexible-pages';
 
 export default () => {
   const featureEnabled = useFeatureFlag({ name: 'customizable_navbar' });