+++ front/app/modules/commercial/customizable_navbar/index.tsx
@@ -1,13 +1,8 @@
 import React from 'react';
 import { ModuleConfiguration } from 'utils/moduleUtils';
-import Tab from './admin/components/Tab';
 import ModuleActive from './admin/components/ModuleActive';
 import PoliciesSubtitle from './admin/components/PoliciesSubtitle';
 
-const CustomNavbarContainer = React.lazy(() => import('./admin/containers'));
-const CustomNavbarSettingsComponent = React.lazy(
-  () => import('./admin/containers/NavigationSettings')
-);
 const NewPageFormComponent = React.lazy(
   () => import('./admin/containers/NewPageForm')
 );
@@ -20,36 +15,22 @@ const EditNavbarItemComponent = React.lazy(
 
 const configuration: ModuleConfiguration = {
   routes: {
-    'admin.settings': [
+    'admin.pages-menu': [
       {
-        path: 'navigation',
-        element: <CustomNavbarContainer />,
-        children: [
-          {
-            // to be changed when refactoring
-            // path: '' should only be used for redirects on
-            // index. Search the codebase for examples
-            path: '',
-            element: <CustomNavbarSettingsComponent />,
-          },
-          {
-            path: 'pages/new',
-            element: <NewPageFormComponent />,
-          },
-          {
-            path: 'pages/edit/:pageId',
-            element: <EditPageComponent />,
-          },
-          {
-            path: 'navbar-items/edit/:navbarItemId',
-            element: <EditNavbarItemComponent />,
-          },
-        ],
+        path: 'pages/new',
+        element: <NewPageFormComponent />,
+      },
+      {
+        path: 'pages/edit/:pageId',
+        element: <EditPageComponent />,
+      },
+      {
+        path: 'navbar-items/edit/:navbarItemId',
+        element: <EditNavbarItemComponent />,
       },
     ],
   },
   outlets: {
-    'app.containers.Admin.settings.tabs': (props) => <Tab {...props} />,
     'app.containers.Admin.settings.customize.Events': (props) => (
       <ModuleActive {...props} />
     ),